requirements.txt
crewai[tools]
litellm
pydantic
python-dotenv
pandas
ddgs
langchain-nvidia-ai-endpoints
langchain-google-genai
langchain-community
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
duckduckgo-search
sentence-transformers
matplotlib